simply scope and rename freeze functions
Rather than enabling user to freeze multiple rows this change
stipulates that only the header row may be frozen. I reason this is
consistent with Comma Chameleon as constrained editor design principle
(in addition to being more straightforward). Calling updateSettings()
twice is a hack but is required due to how the render callbacks
execute. The second call could be used to set `colHeaders` to false,
enabling the user to use their header row as GUI headers